moved classes from test to src
