+++ src/main/java/org/jd/core/v1/impl/printer/PlainTextPrinter.java
@@ -5,7 +5,7 @@
  * copy and modify the code freely for non-commercial purposes.
  */
 
-package org.jd.core.v1.printer;
+package org.jd.core.v1.impl.printer;
 
 
 import org.jd.core.v1.api.printer.Printer;
@@ -19,14 +19,19 @@ public class PlainTextPrinter implements Printer {
     protected int realLineNumber = 0;
     protected String format;
 
-    protected boolean escapeUnicodeCharacters;
+    protected boolean escapeUnicodeCharacters = false;
+    protected boolean showLineNumbers = true;
 
     public PlainTextPrinter() {
-        this.escapeUnicodeCharacters = false;
     }
 
     public PlainTextPrinter(boolean escapeUnicodeCharacters) {
         this.escapeUnicodeCharacters = escapeUnicodeCharacters;
+    }
+
+    public PlainTextPrinter(boolean escapeUnicodeCharacters, boolean showLineNumbers) {
+        this.escapeUnicodeCharacters = escapeUnicodeCharacters;
+        this.showLineNumbers = showLineNumbers;
     }
 
     public void init() {
@@ -123,10 +128,13 @@ public class PlainTextPrinter implements Printer {
     public void endMarker(int type) {}
 
     protected void printLineNumber(int lineNumber) {
-        sb.append("/*");
-        sb.append(String.format(format, ++realLineNumber));
-        sb.append(':');
-        sb.append(String.format(format, lineNumber));
-        sb.append(" */ ");
+    	++realLineNumber;
+    	if (showLineNumbers) {
+	        sb.append("/*");
+	        sb.append(String.format(format, realLineNumber));
+	        sb.append(':');
+	        sb.append(String.format(format, lineNumber));
+	        sb.append(" */ ");
+    	}
     }
 }
