+++ src/test/java/org/jd/core/v1/JavaAnnotationTest.java
@@ -12,10 +12,10 @@ import org.jd.core.v1.api.loader.Loader;
 import org.jd.core.v1.api.printer.Printer;
 import org.jd.core.v1.compiler.CompilerUtil;
 import org.jd.core.v1.compiler.JavaSourceFileObject;
-import org.jd.core.v1.loader.ClassPathLoader;
-import org.jd.core.v1.loader.ZipLoader;
+import org.jd.core.v1.impl.loader.ClassPathLoader;
+import org.jd.core.v1.impl.loader.ZipLoader;
+import org.jd.core.v1.impl.printer.PlainTextPrinter;
 import org.jd.core.v1.model.message.Message;
-import org.jd.core.v1.printer.PlainTextPrinter;
 import org.jd.core.v1.regex.PatternMaker;
 import org.jd.core.v1.service.converter.classfiletojavasyntax.ClassFileToJavaSyntaxProcessor;
 import org.jd.core.v1.service.deserializer.classfile.DeserializeClassFileProcessor;
