+++ src/test/java/org/jd/core/v1/LayoutFragmentProcessorTest.java
@@ -8,9 +8,10 @@
 package org.jd.core.v1;
 
 import junit.framework.TestCase;
-import org.jd.core.v1.loader.ZipLoader;
+
+import org.jd.core.v1.impl.loader.ZipLoader;
+import org.jd.core.v1.impl.printer.PlainTextPrinter;
 import org.jd.core.v1.model.message.Message;
-import org.jd.core.v1.printer.PlainTextPrinter;
 import org.jd.core.v1.service.converter.classfiletojavasyntax.ClassFileToJavaSyntaxProcessor;
 import org.jd.core.v1.service.deserializer.classfile.DeserializeClassFileProcessor;
 import org.jd.core.v1.service.fragmenter.javasyntaxtojavafragment.JavaSyntaxToJavaFragmentProcessor;
