+++ src/test/java/org/jd/core/v1/PopulateBindingsForStaticMethodVisitorTest.java
@@ -8,7 +8,8 @@
 package org.jd.core.v1;
 
 import junit.framework.TestCase;
-import org.jd.core.v1.loader.ClassPathLoader;
+
+import org.jd.core.v1.impl.loader.ClassPathLoader;
 import org.jd.core.v1.model.javasyntax.type.*;
 import org.jd.core.v1.service.converter.classfiletojavasyntax.util.TypeMaker;
 import org.jd.core.v1.service.converter.classfiletojavasyntax.visitor.PopulateBindingsWithTypeArgumentVisitor;
