+++ src/test/java/org/jd/core/v1/WriteTokenTest.java
@@ -7,9 +7,9 @@
 
 package org.jd.core.v1;
 
+import org.jd.core.v1.impl.printer.PlainTextMetaPrinter;
 import org.jd.core.v1.model.message.Message;
 import org.jd.core.v1.model.token.*;
-import org.jd.core.v1.printer.PlainTextMetaPrinter;
 import org.jd.core.v1.service.writer.WriteTokenProcessor;
 import org.jd.core.v1.util.DefaultList;
 import org.junit.Assert;
